style(wishlist): Rename package name to member from user
